pyproject.toml
[project]
name = "test-ldap"
version = "0.1.0"
description = ""
authors = [
    {name = "Artemiy",email = "artemsikor18@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.2,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic[settigns] (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "ldap3 (>=2.9.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
